fix Error

+++ What-s-Learn-Back-End/What'sLearn/ITI.What'sLearn/ITI.WhatsLearnServices/UserTrack/UserTrackService.cs
@@ -34,13 +34,14 @@ namespace ITI.WhatsLearn.Services
         }
         public UserTrack GetByID(int id)
         {
+           
             return UserTrackRepo.GetByID(id);
         }
         public IEnumerable<UserTrackViewModel> GetAll(int pageIndex, int pageSize = 20)
         {
             var query =
                 UserTrackRepo.GetAll();
-            query = query.Skip(pageIndex * pageSize).Take(pageSize);
+            query = query.OrderByDescending(i => i.ID).Skip(pageIndex * pageSize).Take(pageSize);
             return query.ToList().Select(i => i.ToViewModel());
         }
       
@@ -50,7 +51,7 @@ namespace ITI.WhatsLearn.Services
             var query =
                 UserTrackRepo.Get
                     (i => i.IsApproveed==false);
-            query = query.Skip(pageIndex * pageSize).Take(pageSize);
+            query = query.OrderByDescending(i=>i.ID).Skip(pageIndex * pageSize).Take(pageSize);
             return query.ToList().Select(i => i.ToViewModel());
         }
 
@@ -59,7 +60,7 @@ namespace ITI.WhatsLearn.Services
             var query =
                 UserTrackRepo.Get
                     (i => i.IsApproveed==false && i.User.Name==Name);
-            query = query.Skip(pageIndex * pageSize).Take(pageSize);
+            query = query.OrderByDescending(i => i.ID).Skip(pageIndex * pageSize).Take(pageSize);
             return query.ToList().Select(i => i.ToViewModel());
         }
         public IEnumerable<UserTrackViewModel> SearchByTrackName(string TrackName, int pageIndex = 0, int pageSize = 20)
@@ -67,15 +68,28 @@ namespace ITI.WhatsLearn.Services
             var query =
                 UserTrackRepo.Get
                     (i => i.IsApproveed==false&& i.Track.Name==TrackName);
-            query = query.Skip(pageIndex * pageSize).Take(pageSize);
+            query = query.OrderByDescending(i => i.ID).Skip(pageIndex * pageSize).Take(pageSize);
             return query.ToList().Select(i => i.ToViewModel());
         }
 
+        public void Approve(int id)
+        {
 
+            UserTrack u = GetByID(id);
+            if (u != null)
+            {
+                u.IsApproveed = true;
+                unitOfWork.Commit();
+            }
+
+        }
         public void Remove(int id)
         {
-            UserTrackRepo.Remove(UserTrackRepo.GetByID(id));
-            unitOfWork.Commit();
+            if (UserTrackRepo.GetByID(id)!= null)
+            {
+                UserTrackRepo.Remove(UserTrackRepo.GetByID(id));
+                unitOfWork.Commit();
+            }
         }
 
     }
